requirements.txt
streamlit
pytesseract
opencv-python-headless
Pillow
numpy
spacy
scispacy
https://s3-us-west-2.amazonaws.com/ai2-s2-scispacy/releases/v0.5.1/en_core_sci_md-0.5.1.tar.gz
